pyproject.toml
[project]
name = "arc-overlay"
version = "1.3.1"
description = "Real-time OCR-powered gaming overlay for items and quests."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Pillow",
    "numpy",
    "pytesseract",
    "rapidfuzz",
    "pyperclip",
    "pynput",
    "pystray",
    "pyqt6",
    "requests",
    "mss",
    "opencv-python",
]

[project.scripts]
arcoverlay = "arcoverlay:main"

[dependency-groups]
dev = [
    "pyinstaller",
    "cx-Freeze",
    "pytest>=8.0.0",
]
